python/kb/backend/requirements.txt
# Core web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
httpx==0.27.2

# Machine Learning & Embeddings
numpy<2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
openai==1.40.0
FlagEmbedding==1.2.10

# Document parsing
pdfplumber==0.11.0
python-docx==1.1.2

# Database
sqlalchemy==2.0.30
